+++ src/store/mutation-types.js
@@ -6,6 +6,7 @@ export const FETCH_POSTLIST = 'FETCH_POSTLIST'
 export const FETCH_ERROR = 'FETCH_ERROR'
 export const FETCH_SUCCESS = 'FETCH_SUCCESS'
 
+//detail page
 export const FETCH_DETAIL = 'FETCH_DETAIL'
 export const FETCH_DETAIL_SUCCESS = 'FETCH_DETAIL_SUCCESS'
 export const FETCH_DETAIL_ERROR = 'FETCH_DETAIL_ERROR'
@@ -13,8 +14,13 @@ export const UPDATE_DETAIL_ID = 'UPDATE_DETAIL_ID'
 export const UPVOTE_COMMENT = 'UPVOTE_COMMENT'
 export const REPLY_COMMENT = 'REPLY_COMMENT'
 
+//login
 export const UPDATE_USER = 'UPDATE_USER'
 export const FREEZE_LOGINMODAL = 'FREEZE_LOGINMODAL'
+
+//reply
+export const UPDATE_TEMPREPLY = 'UPDATE_TEMPREPLY'
+export const START_REPLY = 'START_REPLY'
 
 export const RESET_PAGE = 'RESET_POSTLIST'
 export const SET_STORAGE = 'SET_STORAGE'